+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -375,6 +375,8 @@ public class SampleSQLBuilder {
 		).put(
 			"releaseDataFactory", new ReleaseDataFactory()
 		).put(
+			"resourcePermissionDataFactory", new ResourcePermissionDataFactory()
+		).put(
 			"socialActivityDataFactory", new SocialActivityDataFactory()
 		).put(
 			"subscriptionDataFactory", new SubscriptionDataFactory()
