+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/class_names.ftl
@@ -1,3 +1,3 @@
 <#list dataFactory.classNameModels as classNameModel>
-	${dataFactory.toInsertSQL(classNameModel)}
+	${resourcePermissionDataFactory.toInsertSQL(classNameModel)}
 </#list>
