+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/commerce_product.ftl
@@ -1,38 +1,38 @@
 <#list dataFactory.newAssetEntryModels() as assetEntryModel>
-	${dataFactory.toInsertSQL(assetEntryModel)}
+	${resourcePermissionDataFactory.toInsertSQL(assetEntryModel)}
 </#list>
 
-${dataFactory.toInsertSQL(commerceDataFactory.newCommerceCatalogModel())}
+${resourcePermissionDataFactory.toInsertSQL(commerceDataFactory.newCommerceCatalogModel())}
 
-${dataFactory.toInsertSQL(dataFactory.newResourcePermission())}
+${resourcePermissionDataFactory.toInsertSQL(resourcePermissionDataFactory.newResourcePermission())}
 
-${dataFactory.toInsertSQL(commerceDataFactory.newCommerceChannelModel())}
+${resourcePermissionDataFactory.toInsertSQL(commerceDataFactory.newCommerceChannelModel())}
 
-${dataFactory.toInsertSQL(commerceDataFactory.newCommerceCurrencyModel())}
+${resourcePermissionDataFactory.toInsertSQL(commerceDataFactory.newCommerceCurrencyModel())}
 
 <#list commerceDataFactory.newCPDefinitionLocalizationModels() as cpDefinitionLocalizationModel>
-	${dataFactory.toInsertSQL(cpDefinitionLocalizationModel)}
+	${resourcePermissionDataFactory.toInsertSQL(cpDefinitionLocalizationModel)}
 </#list>
 
 <#list commerceDataFactory.newCPDefinitionModels() as cpDefinitionModel>
-	${dataFactory.toInsertSQL(cpDefinitionModel)}
+	${resourcePermissionDataFactory.toInsertSQL(cpDefinitionModel)}
 </#list>
 
 <#list commerceDataFactory.newCPFriendlyURLEntryModels() as cpFriendlyURLEntryModel>
-	${dataFactory.toInsertSQL(cpFriendlyURLEntryModel)}
+	${resourcePermissionDataFactory.toInsertSQL(cpFriendlyURLEntryModel)}
 
 	${dataFactory.getCSVWriter("cpFriendlyURLEntry").write(cpFriendlyURLEntryModel.urlTitle + "\n")}
 </#list>
 
 <#list commerceDataFactory.newCPInstanceModels() as cpInstanceModel>
-	${dataFactory.toInsertSQL(cpInstanceModel)}
+	${resourcePermissionDataFactory.toInsertSQL(cpInstanceModel)}
 </#list>
 
 <#list commerceDataFactory.newCProductModels() as cProductModel>
-	${dataFactory.toInsertSQL(cProductModel)}
+	${resourcePermissionDataFactory.toInsertSQL(cProductModel)}
 </#list>
 
-${dataFactory.toInsertSQL(commerceDataFactory.newCPTaxCategoryModel())}
+${resourcePermissionDataFactory.toInsertSQL(commerceDataFactory.newCPTaxCategoryModel())}
 
 <@insertGroup _groupModel=userDataFactory.newCommerceCatalogGroupModel() />
