+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/company.ftl
@@ -1,9 +1,9 @@
 <#assign companyModel = userDataFactory.newCompanyModel() />
 
-${dataFactory.toInsertSQL(companyModel)}
+${resourcePermissionDataFactory.toInsertSQL(companyModel)}
 
-${dataFactory.toInsertSQL(userDataFactory.newAccountModel())}
+${resourcePermissionDataFactory.toInsertSQL(userDataFactory.newAccountModel())}
 
-${dataFactory.toInsertSQL(userDataFactory.newVirtualHostModel())}
+${resourcePermissionDataFactory.toInsertSQL(userDataFactory.newVirtualHostModel())}
 
 ${dataFactory.getCSVWriter("company").write(companyModel.companyId + "\n")}
