+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_dl_file_type.ftl
@@ -1,4 +1,4 @@
-${dataFactory.toInsertSQL(dLDataFactory.newDLFileEntryTypeModel())}
+${resourcePermissionDataFactory.toInsertSQL(dLDataFactory.newDLFileEntryTypeModel())}
 
 <#assign defaultDLDDMStructureModel = dLDataFactory.newDefaultDLDDMStructureModel() />
