+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_user.ftl
@@ -45,7 +45,7 @@
 />
 
 <#list groupIds as groupId>
-	${dataFactory.toInsertSQL(blogDataFactory.newBlogsStatsUserModel(groupId))}
+	${resourcePermissionDataFactory.toInsertSQL(blogDataFactory.newBlogsStatsUserModel(groupId))}
 
-	${dataFactory.toInsertSQL(messageBoardDataFactory.newMBStatsUserModel(groupId))}
+	${resourcePermissionDataFactory.toInsertSQL(messageBoardDataFactory.newMBStatsUserModel(groupId))}
 </#list>
