+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/release.ftl
@@ -1,5 +1,5 @@
 <#assign releaseModels = releaseDataFactory.newReleaseModels() />
 
 <#list releaseModels as releaseModel>
-	${dataFactory.toInsertSQL(releaseModel)}
+	${resourcePermissionDataFactory.toInsertSQL(releaseModel)}
 </#list>
